better user ids and a message bus limit

+++ src/domain/types.rs
@@ -8,5 +8,5 @@ use tokio_tungstenite::tungstenite::Message;
 
 use super::{room::Room, user::User};
 
-pub type PeerMap = Arc<Mutex<HashMap<u64, (Arc<Mutex<User>>, UnboundedSender<Message>)>>>;
+pub type PeerMap = Arc<Mutex<HashMap<u128, (Arc<Mutex<User>>, UnboundedSender<Message>)>>>;
 pub type RoomMap = Arc<Mutex<HashMap<u64, Room>>>;
